Move substate_mut logic into track
